Django API/atbapi/requirements.txt
absl-py==2.3.1
asgiref==3.9.2
astunparse==1.6.3
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.6
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg==1.21.11
flatbuffers==25.9.23
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.3
namex==0.1.0
numpy==2.3.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
protobuf==6.32.1
Pygments==2.19.2
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
